Relocated DashboardNav to dashboard/nav/components for better feature-based organization
