+++ src/app/dashboard/layout.tsx
@@ -4,7 +4,7 @@ import { FC, ReactNode, useEffect } from 'react';
 import { useAuth } from '@/context/AuthContext';
 import { useRouter } from 'next/navigation';
 
-import DashboardNav from '@/features/dashboard/DashboardNav';
+import DashboardNav from '@/features/dashboard/nav/components/DashboardNav';
 
 interface DashboardLayoutProps {
   children: ReactNode;
